LayoutTests/http/tests/eventloop/documents-with-different-protocols-do-not-share-event-loop-expected.txt

This tests the order by which tasks are scheduled across documents that are not similar origins.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.hasSameEventLoopAs(frame1.contentWindow) is true
PASS internals.hasSameEventLoopAs(frame2.contentWindow) is false
PASS successfullyParsed is true

TEST COMPLETE
